# backend/specs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for archery statistics tracking",
        "title": "Arch Stats API",
        "contact": {},
        "version": "v0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/archer/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a single archer profile by UUID primary key identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Archers"
                ],
                "summary": "Get Archer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Archer UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ArcherRead"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove an archer profile by UUID primary key identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Archers"
                ],
                "summary": "Delete Archer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Archer UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/session/slot": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Assigns an archer to a target slot within an open session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slots"
                ],
                "summary": "Join Session",
                "parameters": [
                    {
                        "description": "Slot join parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SlotJoinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SlotJoinResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/session/slot/archer/{archer_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns active slot assignment for the given archer (open session and is_shooting = true)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slots"
                ],
                "summary": "Get Archer Current Slot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Archer UUID",
                        "name": "archer_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/FullSlotInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/session/slot/leave/{slot_id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deactivates an active slot assignment (stop shooting in the session)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slots"
                ],
                "summary": "Leave Session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slot UUID",
                        "name": "slot_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/session/slot/re-join/{slot_id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Re-activates a previously inactive slot assignment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slots"
                ],
                "summary": "Re Join Session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slot UUID",
                        "name": "slot_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SlotJoinResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/session/slot/{slot_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns active slot assignment details by slot UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slots"
                ],
                "summary": "Get Slot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slot UUID",
                        "name": "slot_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/FullSlotInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ArcherRead": {
            "type": "object",
            "required": [
                "archer_id",
                "bowstyle",
                "created_at",
                "date_of_birth",
                "draw_weight",
                "email",
                "first_name",
                "gender",
                "google_subject",
                "last_login_at",
                "last_name"
            ],
            "properties": {
                "archer_id": {
                    "type": "string"
                },
                "bowstyle": {
                    "$ref": "#/definitions/Bowstyle"
                },
                "club_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "draw_weight": {
                    "type": "number"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/Gender"
                },
                "google_picture_url": {
                    "type": "string"
                },
                "google_subject": {
                    "type": "string"
                },
                "last_login_at": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
...
        "Bowstyle": {
            "type": "string",
            "enum": [
                "recurve",
                "compound",
                "barebow",
                "longbow"
            ],
            "x-enum-varnames": [
                "BowstyleRecurve",
                "BowstyleCompound",
                "BowstyleBarebow",
                "BowstyleLongbow"
            ]
        },
        "ErrorResponse": {
            "type": "object",
            "required": [
                "detail"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                }
            }
        },
...
        "FaceType": {
            "type": "string",
            "enum": [
                "wa_40cm_full",
                "wa_60cm_full",
                "wa_80cm_full",
                "wa_122cm_full",
                "wa_40cm_6rings",
                "wa_60cm_6rings",
                "wa_80cm_6rings",
                "wa_122cm_6rings",
                "wa_40cm_triple_vertical",
                "wa_60cm_triple_triangular",
                "none"
            ],
            "x-enum-varnames": [
                "FaceTypeWA40Full",
                "FaceTypeWA60Full",
                "FaceTypeWA80Full",
                "FaceTypeWA122Full",
                "FaceTypeWA406Rings",
                "FaceTypeWA606Rings",
                "FaceTypeWA806Rings",
                "FaceTypeWA1226Rings",
                "FaceTypeWA40TripleVertical",
                "FaceTypeWA60TripleTriangular",
                "FaceTypeNone"
            ]
        },
        "FullSlotInfo": {
            "type": "object",
            "required": [
                "archer_id",
                "bowstyle",
                "created_at",
                "distance",
                "draw_weight",
                "face_type",
                "interval_seconds",
                "is_shooting",
                "lane",
                "session_id",
                "slot",
                "slot_id",
                "slot_letter",
                "target_id"
            ],
            "properties": {
                "archer_id": {
                    "type": "string"
                },
                "bowstyle": {
                    "$ref": "#/definitions/Bowstyle"
                },
                "club_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "distance": {
                    "type": "integer"
                },
                "draw_weight": {
                    "type": "number"
                },
                "face_type": {
                    "$ref": "#/definitions/FaceType"
                },
                "interval_seconds": {
                    "type": "integer"
                },
                "is_shooting": {
                    "type": "boolean"
                },
                "lane": {
                    "type": "integer"
                },
                "session_id": {
                    "type": "string"
                },
                "shot_per_round": {
                    "type": "integer"
                },
                "slot": {
                    "type": "string"
                },
                "slot_id": {
                    "type": "string"
                },
                "slot_letter": {
                    "$ref": "#/definitions/SlotLetter"
                },
                "target_id": {
                    "type": "string"
                }
            }
        },
        "Gender": {
            "type": "string",
            "enum": [
                "male",
                "female",
                "non_binary",
                "other",
                "unspecified"
            ],
            "x-enum-varnames": [
                "GenderMale",
                "GenderFemale",
                "GenderNonBinary",
                "GenderOther",
                "GenderUnspecified"
            ]
        },
...
        "SlotJoinRequest": {
            "type": "object",
            "required": [
                "archer_id",
                "bowstyle",
                "distance",
                "draw_weight",
                "face_type",
                "interval_seconds",
                "is_shooting",
                "session_id"
            ],
            "properties": {
                "archer_id": {
                    "type": "string"
                },
                "bowstyle": {
                    "$ref": "#/definitions/Bowstyle"
                },
                "club_id": {
                    "type": "string"
                },
                "distance": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                },
                "draw_weight": {
                    "type": "number",
                    "maximum": 200
                },
                "face_type": {
                    "$ref": "#/definitions/FaceType"
                },
                "interval_seconds": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                },
                "is_shooting": {
                    "type": "boolean"
                },
                "session_id": {
                    "type": "string"
                },
                "shot_per_round": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 3
                }
            }
        },
        "SlotJoinResponse": {
            "type": "object",
            "required": [
                "slot",
                "slot_id"
            ],
            "properties": {
                "slot": {
                    "type": "string"
                },
                "slot_id": {
                    "type": "string"
                }
            }
        },
        "SlotLetter": {
            "type": "string",
            "enum": [
                "A",
                "B",
                "C",
                "D"
            ],
            "x-enum-varnames": [
                "SlotLetterA",
                "SlotLetterB",
                "SlotLetterC",
                "SlotLetterD"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}